removido RelatorioPendenciasController
